// app/types.ts
export interface Product {
  id: number;
  imageSrc: string;
  media: string[];
  title: string;
  description: string;
  pastDescription: string;
  badge: string;
  date: string;
  endDate:string,
  location: string;
  paymentLink: string;
  // price: number;
}

// export interface Event {
//   id: number;
//   imageSrc: string;
//   title: string;
//   description: string;
//   link: string;
// }